Fixed beats on oscillator second boundaries
Was arbitrarily restarting period on one-second marks. Fixed.

Made the clock a little more random-access.

Stick with pure u32 longer in the pipeline and derive f32 as needed from them.

+++ src/backend/orchestrator.rs
@@ -16,7 +16,7 @@ pub struct Orchestrator {
 impl Orchestrator {
     pub fn new(sample_rate: u32) -> Orchestrator {
         Orchestrator {
-            clock: Clock::new(sample_rate as f32),
+            clock: Clock::new(sample_rate),
             master_mixer: Rc::new(RefCell::new(Mixer::new())),
             devices: Vec::new(),
         }
